backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pillow==10.4.0
pydantic==2.8.2
requests==2.32.3
python-multipart==0.0.9